webkitgtk-2.34.6+wayfarer-3

+++ Tools/Scripts/libraries/webkitcorepy/webkitcorepy/task_pool.py
@@ -27,20 +27,25 @@ import multiprocessing
 import signal
 import sys
 
+if sys.version_info < (3, 0):
+    import Queue
+else:
+    import queue as Queue
+
 from webkitcorepy import OutputCapture, Timeout, log
 
 
-class Message(object):
+class _Message(object):
     def __init__(self, who=None):
-        self.who = who or Process.name
+        self.who = who or _Process.name
 
     def __call__(self, caller):
         NotImplemented()
 
 
-class Task(Message):
+class _Task(_Message):
     def __init__(self, function, id, *args, **kwargs):
-        super(Task, self).__init__()
+        super(_Task, self).__init__()
 
         self.function = function
         self.id = id
@@ -51,9 +56,9 @@ class Task(Message):
         return self.function(*self.args, **self.kwargs)
 
 
-class Result(Message):
+class _Result(_Message):
     def __init__(self, value, id):
-        super(Result, self).__init__()
+        super(_Result, self).__init__()
         self.value = value
         self.id = id
 
@@ -63,21 +68,21 @@ class Result(Message):
         return self.value
 
 
-class Log(Message):
+class _Log(_Message):
     def __init__(self, record):
-        super(Log, self).__init__()
+        super(_Log, self).__init__()
         self.record = record
 
     def __call__(self, caller):
         logging.getLogger(self.record.name).log(self.record.levelno, '{} {}'.format(self.who, self.record.getMessage()))
 
 
-class Print(Message):
+class _Print(_Message):
     stdout = 1
     stderr = 2
 
     def __init__(self, lines, stream=stdout):
-        super(Print, self).__init__()
+        super(_Print, self).__init__()
         self.lines = lines
         self.stream = stream
 
@@ -90,32 +95,32 @@ class Print(Message):
             stream.write('{}\n'.format(line))
 
 
-class State(Message):
+class _State(_Message):
     STARTING, STOPPING = 1, 0
     STATES = [STARTING, STOPPING]
 
     def __init__(self, state):
-        super(State, self).__init__()
+        super(_State, self).__init__()
         self.state = state
 
     def __call__(self, caller):
         log.info('{} {}'.format(
             self.who, {
-                State.STARTING: 'starting',
-                State.STOPPING: 'stopping',
+                self.STARTING: 'starting',
+                self.STOPPING: 'stopping',
             }.get(self.state, self.state),
         ))
         if caller:
             caller._started += {
-                State.STARTING: 1,
-                State.STOPPING: -1,
+                self.STARTING: 1,
+                self.STOPPING: -1,
             }.get(self.state, 0)
         return self.state
 
 
-class ChildException(Message):
+class _ChildException(_Message):
     def __init__(self, exc_info=None):
-        super(ChildException, self).__init__()
+        super(_ChildException, self).__init__()
         self.exc_info = exc_info or sys.exc_info()
 
     def __call__(self, caller):
@@ -128,7 +133,7 @@ class ChildException(Message):
         reraise(*self.exc_info)
 
 
-class BiDirectionalQueue(object):
+class _BiDirectionalQueue(object):
     def __init__(self, outgoing=None, incoming=None):
         self.outgoing = outgoing or multiprocessing.Queue()
         self.incoming = incoming or multiprocessing.Queue()
@@ -146,18 +151,39 @@ class BiDirectionalQueue(object):
                 return self.incoming.get(timeout=difference)
             return self.incoming.get()
 
+    def close(self):
+        with OutputCapture():
+            self.outgoing.close()
+            self.incoming.close()
+            self.outgoing.join_thread()
+            self.incoming.join_thread()
 
-class Process(object):
+
+class _DummyQueue(object):
+    def send(self, object):
+        if isinstance(object, _Message):
+            object(None)
+        return True
+
+    def receive(self, blocking=True):
+        pass
+
+    def close(self):
+        pass
+
+
+class _Process(object):
     name = None
     working = False
+    queue = None
 
     class LogHandler(logging.Handler):
         def __init__(self, queue, **kwargs):
             self._queue = queue
-            super(Process.LogHandler, self).__init__(**kwargs)
+            super(_Process.LogHandler, self).__init__(**kwargs)
 
         def emit(self, record):
-            self._queue.send(Log(record))
+            self._queue.send(_Log(record))
 
     class Stream(io.IOBase):
         def __init__(self, handle, queue):
@@ -169,7 +195,7 @@ class Process(object):
 
         def flush(self):
             if self.cache is not None:
-                self._queue.send(Print(lines=[self.cache], stream=self.handle))
+                self._queue.send(_Print(lines=[self.cache], stream=self.handle))
                 self.cache = None
 
         def writelines(self, lines):
@@ -185,7 +211,7 @@ class Process(object):
                 elif c not in ['\r', '\0']:
                     self.cache = c if self.cache is None else (self.cache + c)
             if to_be_printed:
-                self._queue.send(Print(lines=to_be_printed, stream=self.handle))
+                self._queue.send(_Print(lines=to_be_printed, stream=self.handle))
             return len(data)
 
         @property
@@ -227,97 +253,155 @@ class Process(object):
 
     @classmethod
     def handler(cls, value, _):
-        if value == getattr(signal, 'SIGTERM'):
+        if value in (getattr(signal, 'SIGTERM'), getattr(signal, 'SIGINT')):
             cls.working = False
 
     @classmethod
-    def main(cls, name, setup, queue, teardown):
+    def main(cls, name, loglevel, setup, setupargs, setupkwargs, queue, teardown, teardownargs, teardownkwargs):
         from tblib import pickling_support
 
         cls.name = name
         cls.working = True
 
+        setupargs = setupargs or []
+        setupkwargs = setupkwargs or {}
+
+        teardownargs = teardownargs or []
+        teardownkwargs = teardownkwargs or {}
+
         if getattr(signal, 'SIGTERM'):
             signal.signal(signal.SIGTERM, cls.handler)
+        if getattr(signal, 'SIGINT'):
+            signal.signal(signal.SIGINT, cls.handler)
 
         logger = logging.getLogger()
         for handler in logger.handlers:
             logger.removeHandler(handler)
         logger.addHandler(cls.LogHandler(queue))
+        logger.setLevel(loglevel)
 
-        queue.send(State(State.STARTING))
+        cls.queue = queue
+        queue.send(_State(_State.STARTING))
 
-        with OutputCapture.ReplaceSysStream('stderr', cls.Stream(Print.stderr, queue)), OutputCapture.ReplaceSysStream('stdout', cls.Stream(Print.stdout, queue)):
+        with OutputCapture.ReplaceSysStream('stderr', cls.Stream(_Print.stderr, queue)), OutputCapture.ReplaceSysStream('stdout', cls.Stream(_Print.stdout, queue)):
             try:
                 pickling_support.install()
                 if setup:
-                    setup()
+                    setup(*setupargs, **setupkwargs)
 
                 while cls.working:
                     task = queue.receive()
                     if not task:
                         break
-                    queue.send(Result(value=task(None), id=task.id))
+                    queue.send(_Result(value=task(None), id=task.id))
 
             except BaseException:
                 typ, exception, traceback = sys.exc_info()
-                queue.send(ChildException(exc_info=(
+                queue.send(_ChildException(exc_info=(
                     typ, typ('{} (from {})'.format(str(exception), name)), traceback,
                 )))
 
             finally:
                 if teardown:
-                    teardown()
+                    teardown(*teardownargs, **teardownkwargs)
                 sys.stdout.flush()
                 sys.stderr.flush()
-                queue.send(State(State.STOPPING))
+                queue.send(_State(_State.STOPPING))
+                cls.queue.close()
+                cls.queue = None
 
 
 class TaskPool(object):
+    Message = _Message
+    Task = _Task
+    Result = _Result
+    Log = _Log
+    Print = _Print
+    State = _State
+    ChildException = _ChildException
+    BiDirectionalQueue = _BiDirectionalQueue
+    Process = _Process
+
 
     class Exception(RuntimeError):
         pass
 
-    def __init__(self, workers=1, name=None, setup=None, teardown=None, grace_period=5, block_size=1000):
+    def __init__(
+        self, workers=1, name=None, setup=None, teardown=None, grace_period=5, block_size=1000,
+        setupargs=None, setupkwargs=None,
+        teardownargs=None, teardownkwargs=None,
+        force_fork=False,
+    ):
         # Ensure tblib is installed before creating child processes
         import tblib
 
         name = name or 'worker'
-        if name == Process.name:
+        if name == self.Process.name:
             raise ValueError("Parent process is already named {}".format(name))
+        self.name = name
 
         if workers < 1:
             raise ValueError('TaskPool requires positive number of workers')
 
-        self.queue = BiDirectionalQueue()
+        self.queue = None
+        self.workers = []
 
-        self.workers = [multiprocessing.Process(
-            target=Process.main,
-            args=('{}/{}'.format(name, count), setup, BiDirectionalQueue(
-                outgoing=self.queue.incoming,
-                incoming=self.queue.outgoing,
-            ), teardown),
-        ) for count in range(workers)]
+        self._setup_args = (setup, setupargs or [], setupkwargs or {})
+        self._teardown_args = (teardown, teardownargs or [], teardownkwargs or {})
+        self._num_workers = int(workers)
+
         self._started = 0
 
         self.callbacks = {}
         self._id_count = 0
         self.grace_period = grace_period
         self.block_size = block_size
+        self.force_fork = force_fork
+
+        if not self.force_fork and self._num_workers == 1 and TaskPool.Process.queue:
+            raise ValueError('Illegal single-process TaskPool nesting')
 
     def __enter__(self):
+        if not self.force_fork and self._num_workers == 1:
+            TaskPool.Process.queue = _DummyQueue()
+            TaskPool.Process.name = TaskPool.Process.name or '{}/0'.format(self.name)
+            if self._setup_args[0]:
+                self._setup_args[0](*self._setup_args[1], **self._setup_args[2])
+            TaskPool.Process.working = True
+            return self
+
+        from mock import patch
+
+        self.queue = self.BiDirectionalQueue()
+        self.workers = [multiprocessing.Process(
+            target=self.Process.main,
+            args=(
+                '{}/{}'.format(self.name, count), logging.getLogger().getEffectiveLevel(),
+                self._setup_args[0], self._setup_args[1], self._setup_args[2],
+                self.BiDirectionalQueue(outgoing=self.queue.incoming, incoming=self.queue.outgoing),
+                self._teardown_args[0], self._teardown_args[1], self._teardown_args[2],
+            ),
+        ) for count in range(self._num_workers)]
+
         with Timeout(seconds=10, patch=False, handler=self.Exception('Failed to start all workers')):
             for worker in self.workers:
                 worker.start()
             while self._started < len(self.workers):
                 self.queue.receive()(self)
+
         return self
 
     def do(self, function, *args, **kwargs):
         callback = kwargs.pop('callback', None)
+        if not self.queue:
+            result = function(*args, **kwargs)
+            if callback:
+                callback(result)
+            return
+
         if callback:
             self.callbacks[self._id_count] = callback
-        self.queue.send(Task(function, self._id_count, *args, **kwargs))
+        self.queue.send(self.Task(function, self._id_count, *args, **kwargs))
         self._id_count += 1
 
         # For every block of tasks passed to our workers, we need consume messages so we don't get deadlocked
@@ -325,10 +409,13 @@ class TaskPool(object):
             while True:
                 try:
                     self.queue.receive(blocking=False)(self)
-                except Exception:
+                except Queue.Empty:
                     break
 
     def wait(self):
+        if not self.queue:
+            return
+
         for _ in self.workers:
             self.queue.send(None)
 
@@ -336,6 +423,16 @@ class TaskPool(object):
             self.queue.receive()(self)
 
     def __exit__(self, *args, **kwargs):
+        if not self.queue:
+            TaskPool.Process.working = False
+            try:
+                if self._teardown_args[0]:
+                    self._teardown_args[0](*self._teardown_args[1], **self._teardown_args[2])
+            finally:
+                TaskPool.Process.queue = None
+                TaskPool.Process.name = None
+            return
+
         from six import reraise
         try:
             inflight = sys.exc_info()
@@ -361,7 +458,15 @@ class TaskPool(object):
 
                 if sys.version_info >= (3, 7):
                     worker.kill()
-                elif hasattr(signal, 'SIGKILL'):
-                    os.kill(worker.pid, signal.SIGKILL)
+                # With python2 killing directly the workers causes the queue to hang on close() <https://webkit.org/b/227715>
+                elif hasattr(signal, 'SIGKILL') and sys.version_info.major > 2:
+                    try:
+                        os.kill(worker.pid, signal.SIGKILL)
+                    except OSError as e:
+                        log.warn('Failed to terminate worker ' + str(worker.pid) + ' with error ' + str(e))
                 else:
                     worker.terminate()
+
+            self.queue.close()
+            self.queue = None
+            self.workers = []
